+++ src/directives/geojson.js
@@ -1,13 +1,12 @@
 angular.module("leaflet-directive")
-.directive('geojson', function ($log, $rootScope, leafletData, leafletHelpers,
-    leafletWatchHelpers, leafletDirectiveControlsHelpers,leafletIterators,
-    leafletGeoJsonEvents) {
-
+.directive('geojson', function (leafletLogger, $rootScope, leafletData, leafletHelpers,
+    leafletWatchHelpers, leafletDirectiveControlsHelpers,leafletIterators, leafletGeoJsonEvents) {
     var _maybeWatch = leafletWatchHelpers.maybeWatch,
         _watchOptions = leafletHelpers.watchOptions,
         _extendDirectiveControls = leafletDirectiveControlsHelpers.extend,
         hlp = leafletHelpers,
         $it = leafletIterators;
+        // $log = leafletLogger;
 
     return {
         restrict: "A",
@@ -35,7 +34,7 @@ angular.module("leaflet-directive")
                                 layer.bindLabel(feature.properties.description);
                             }
 
-                            leafletGeoJsonEvents.bindEvents(layer, null, feature,
+                            leafletGeoJsonEvents.bindEvents(attrs.id, layer, null, feature,
                                 leafletScope, maybeName,
                                 {resetStyleOnMouseout: geojson.resetStyleOnMouseout,
                                 mapId: attrs.id});