+++ src/directives/paths.js
@@ -1,4 +1,5 @@
-angular.module("leaflet-directive").directive('paths', function ($log, $q, leafletData, leafletMapDefaults, leafletHelpers, leafletPathsHelpers, leafletEvents) {
+angular.module("leaflet-directive").directive('paths', function (leafletLogger, $q, leafletData, leafletMapDefaults, leafletHelpers, leafletPathsHelpers, leafletPathEvents) {
+    var $log = leafletLogger;
     return {
         restrict: "A",
         scope: false,
@@ -12,7 +13,7 @@ angular.module("leaflet-directive").directive('paths', function ($log, $q, leafl
                 leafletScope  = mapController.getLeafletScope(),
                 paths     = leafletScope.paths,
                 createPath = leafletPathsHelpers.createPath,
-                bindPathEvents = leafletEvents.bindPathEvents,
+                bindPathEvents = leafletPathEvents.bindPathEvents,
                 setPathOptions = leafletPathsHelpers.setPathOptions;
 
             mapController.getMap().then(function(map) {
@@ -138,7 +139,7 @@ angular.module("leaflet-directive").directive('paths', function ($log, $q, leafl
                                     }
                                 }
 
-                                bindPathEvents(newPath, newName, pathData, leafletScope);
+                                bindPathEvents(attrs.id, newPath, newName, pathData, leafletScope);
                             }
                         }
                     });